BDOG-2805 add v2 endpoint that uses new slack api and is protected by internal-auth
