+++ build.sbt
@@ -8,3 +8,4 @@ lazy val microservice = Project("slack-notifications", file("."))
   .settings(libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test)
   .settings(resolvers += Resolver.jcenterRepo)
   .settings(scalaVersion := "2.13.10")
+  .settings(scalacOptions += "-Wconf:src=routes/.*:s")
