+++ conf/application.conf
@@ -53,8 +53,10 @@ microservice {
 }
 
 slack {
-    webhookUrl = PLACEHOLDER
-    notification.enabled = true
+  webhookUrl = PLACEHOLDER
+  notification.enabled = true
+  apiUrl = "https://slack.com/api"
+  botToken = PLACEHOLDER
 }
 
 alerts {
