+++ project/AppDependencies.scala
@@ -7,8 +7,9 @@ object AppDependencies {
   val bootstrapPlayVersion = "7.22.0"
 
   val compile = Seq(
-    "uk.gov.hmrc"           %% "bootstrap-backend-play-28" % bootstrapPlayVersion,
-    "org.typelevel"         %% "cats-core"                 % "2.10.0",
+    "uk.gov.hmrc"           %% "bootstrap-backend-play-28"    % bootstrapPlayVersion,
+    "uk.gov.hmrc"           %% "internal-auth-client-play-28" % "1.6.0",
+    "org.typelevel"         %% "cats-core"                    % "2.10.0",
     ws,
     ehcache
   )
